+++ packages/aeson/src/Conferer/Source/Aeson.hs
@@ -22,10 +22,14 @@ import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as L
 import Data.List (intersperse)
 import System.Directory (doesFileExist)
+import Control.Exception
 
 import Conferer.Source.Files
 import qualified Conferer.Source.Null as Null
 import Conferer.Source
+import Control.Monad (guard)
+import Control.Exception (throwIO)
+import Conferer.Key (isValidKeyFragment)
 
 -- | 'Source' that read a config file as json and uses that value in a way that
 -- makes sense for Conferer but doesn't respect json perfectly.
@@ -70,9 +74,20 @@ fromFilePath' fileToParse = do
         Nothing ->
           error $ "Failed to decode json file '" ++ fileToParse ++ "'"
         Just v -> do
-          return $ fromValue v
+          case validateJsonKeys v of
+            [] ->
+              return $ fromValue v
+            errors ->
+              throwIO $ JsonHasInvalidKeysError fileToParse errors
     else do
       return $ Null.empty
+
+-- | Exception thrown from 'fromFilePath' when the json in the
+-- parsed file has incorrect keys
+data JsonHasInvalidKeysError =
+  JsonHasInvalidKeysError FilePath [RawKey] deriving (Eq, Show)
+
+instance Exception JsonHasInvalidKeysError
 
 -- | Create a 'Source' from a json value, never fails.
 fromValue :: Value -> Source
@@ -126,9 +141,16 @@ listKeysInJSON = go ""
   go :: Key -> Value -> [Key]
   go key value =
     case (unconsKey key, value) of
-      (_, Object o) -> do
-        (k, v) <- HashMap.toList o
-        go (key /. fromText k) v
+      (_, Object o) ->
+        let
+          self =
+            case valueToText <$> HashMap.lookup "_self" o of
+              Just _ -> [key]
+              Nothing -> []
+        in self ++ do
+          (k, v) <- HashMap.toList o
+          guard $ isValidKeyFragment k
+          go (key /. fromText k) v
       (_, Array as) -> do
         (index :: Integer, v) <- zip [0..] $ Vector.toList as
         go (key /. mkKey (show index)) v
@@ -138,7 +160,9 @@ listKeysInJSON = go ""
 -- | Turn json 'Value' into 'Text' to return that key
 valueToText :: Value -> Maybe Text
 valueToText (String t) = Just t
-valueToText (Object _o) = Nothing
+valueToText (Object o) = do
+  selfValue <- HashMap.lookup "_self" o
+  valueToText selfValue
 valueToText (Array _as) = Nothing
 valueToText (Number n) = Just $ Text.decodeUtf8 $ L.toStrict $ encode $ Number n
 valueToText (Bool b) = Just $ boolToString b
@@ -154,3 +178,31 @@ resultToMaybe :: Result a -> Maybe a
 resultToMaybe (Error _) = Nothing
 resultToMaybe (Success a) = Just a
 
+type RawKey = [Text]
+
+-- | Validates that a json has the correct format for keys,
+-- since Conferer 'Key's are pretty restricted.
+--
+-- The Source will work with incorrect keys, it will ignore
+-- them which is not intuitive for the user.
+validateJsonKeys :: Value -> [RawKey]
+validateJsonKeys = go []
+  where
+  go :: RawKey -> Value -> [RawKey]
+  go key value =
+    case value of
+      Object o ->
+        let
+          wrongKeys =
+            fmap (\t -> key ++ [t])
+            . filter (not . (== "_self"))
+            . filter (not . isValidKeyFragment)
+            . HashMap.keys
+            $ o
+        in wrongKeys ++ do
+        (k, v) <- HashMap.toList o
+        go (key ++ [k]) v
+      Array as -> do
+        (index :: Integer, v) <- zip [0..] $ Vector.toList as
+        go (key ++ [Text.pack $ show index]) v
+      _ -> []